+++ src/Aws/Kinesis.hs
@@ -1,9 +1,25 @@
-{-# LANGUAGE OverloadedStrings #-}
+-- Copyright (c) 2013-2015 PivotCloud, Inc.
+--
+-- Aws.Kinesis
+--
+-- Please feel free to contact us at licensing@pivotmail.com with any
+-- contributions, additions, or other feedback; we would love to hear from
+-- you.
+--
+-- Licensed under the Apache License, Version 2.0 (the "License"); you may
+-- not use this file except in compliance with the License. You may obtain a
+-- copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+--
+-- Unless required by applicable law or agreed to in writing, software
+-- distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+-- WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+-- License for the specific language governing permissions and limitations
+-- under the License.
 
 -- |
 -- Module: Aws.Kinesis
--- Copyright: Copyright © 2014 AlephCloud Systems, Inc.
--- License: MIT
+-- Copyright: Copyright (c) 2013-2015 PivotCloud, Inc.
+-- license: Apache License, Version 2.0
 -- Maintainer: Lars Kuhtz <lars@alephcloud.com>
 -- Stability: experimental
 --
@@ -30,7 +46,9 @@
 --
 -- In order to run the example you must replace @"access-key-id"@ and
 -- @"secret-access-key"@ with the respective values for your AWS account.
---
+
+{-# LANGUAGE OverloadedStrings #-}
+
 module Aws.Kinesis
 ( module Aws.Kinesis.Core
 , module Aws.Kinesis.Types
@@ -42,6 +60,7 @@ module Aws.Kinesis
 , module Aws.Kinesis.Commands.ListStreams
 , module Aws.Kinesis.Commands.MergeShards
 , module Aws.Kinesis.Commands.PutRecord
+, module Aws.Kinesis.Commands.PutRecords
 , module Aws.Kinesis.Commands.SplitShard
 ) where
 
@@ -55,4 +74,5 @@ import Aws.Kinesis.Commands.GetShardIterator
 import Aws.Kinesis.Commands.ListStreams
 import Aws.Kinesis.Commands.MergeShards
 import Aws.Kinesis.Commands.PutRecord
+import Aws.Kinesis.Commands.PutRecords
 import Aws.Kinesis.Commands.SplitShard